Why did the chicken cross the road?
Because the chicken was actually a markdown file that had been committed to crossing the road. All he needed was that final push. Yes this commit was useless, but I wanted to be first... and see how this new IDE worked... and because I always like to start my srs bsns with a light joke.